Bug #650165 [iSharkL2][BringUpTask] Open camera in 7.0_k44 branch
[root cause  ] open camera and fix compiled error
[changes     ] libcamera folder
[side effects] no
[self test   ]<board, Test cases summary>:hal compile ok
[reviewers   ] jianping.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ sensor/sensor_drv/classic/Other/c2390/sensor_c2390_mipi_raw.c
@@ -1545,7 +1545,7 @@ static uint32_t c2390_write_exposure(SENSOR_HW_HANDLE handle,uint32_t param)
 	return ret_value;
 }
 
-static uint32_t c2390_write_exposure_ex(SENSOR_HW_HANDLE handle,uint32_t param)
+static uint32_t c2390_write_exposure_ex(SENSOR_HW_HANDLE handle,unsigned long param)
 {
 	uint32_t ret_value = SENSOR_SUCCESS;
 	uint16_t exposure_line = 0x00;
